Relocated all CLI commands into cli package
