+++ pyproject.toml
@@ -65,6 +65,7 @@ where = ["src"]
 [tool.pytest.ini_options]
 pythonpath = "src/"
 addopts = ["--import-mode=importlib"]
+testpaths = ["tests"]
 markers = []
 
 [dependency-groups]
