+++ tests/rtf_to_text/test_cli.py
@@ -7,7 +7,8 @@ from unittest.mock import patch
 import pytest
 from click.testing import CliRunner
 
-from kp_analysis_toolkit.rtf_to_text.cli import get_input_file, process_command_line
+from kp_analysis_toolkit.cli.commands.rtf_to_text import process_command_line
+from kp_analysis_toolkit.cli.common.file_selection import get_input_file
 
 
 class TestGetInputFile:
@@ -22,7 +23,12 @@ class TestGetInputFile:
         """Test error when no RTF files are found."""
         with tempfile.TemporaryDirectory() as tmpdir:  # noqa: SIM117
             with pytest.raises(ValueError, match="No RTF files found"):
-                get_input_file(None, tmpdir)
+                get_input_file(
+                    None,
+                    tmpdir,
+                    file_pattern="*.rtf",
+                    file_type_description="RTF",
+                )
 
     def test_get_input_file_single_rtf_file(self) -> None:
         """Test automatic selection when only one RTF file exists."""
@@ -31,7 +37,12 @@ class TestGetInputFile:
             rtf_file = tmpdir_path / "test.rtf"
             rtf_file.write_text("{\rtf1 test}")
 
-            result = get_input_file(None, tmpdir)
+            result = get_input_file(
+                None,
+                tmpdir,
+                file_pattern="*.rtf",
+                file_type_description="RTF",
+            )
             # Resolve both paths to handle Windows short/long path differences
             assert result.resolve() == rtf_file.resolve()
 
@@ -45,13 +56,40 @@ class TestGetInputFile:
             rtf_file2.write_text("{\rtf1 test2}")
 
             with patch("builtins.input", return_value="1"):
-                result = get_input_file(None, tmpdir)
+                result = get_input_file(
+                    None,
+                    tmpdir,
+                    file_pattern="*.rtf",
+                    file_type_description="RTF",
+                )
 
             # Result should be one of the files (order may vary based on filesystem)
             # Resolve paths to handle Windows short/long path differences
             resolved_result = result.resolve()
             resolved_files = [rtf_file1.resolve(), rtf_file2.resolve()]
             assert resolved_result in resolved_files
+
+    def test_get_input_file_process_all_option(self) -> None:
+        """Test the 'process all files' option functionality."""
+        with tempfile.TemporaryDirectory() as tmpdir:
+            tmpdir_path = Path(tmpdir)
+            rtf_file1 = tmpdir_path / "test1.rtf"
+            rtf_file2 = tmpdir_path / "test2.rtf"
+            rtf_file1.write_text("{\rtf1 test1}")
+            rtf_file2.write_text("{\rtf1 test2}")
+
+            # User chooses option 3 (process all files)
+            with patch("builtins.input", return_value="3"):
+                result = get_input_file(
+                    None,
+                    tmpdir,
+                    file_pattern="*.rtf",
+                    file_type_description="RTF",
+                    include_process_all_option=True,
+                )
+
+            # Should return None to indicate "process all files"
+            assert result is None
 
 
 class TestProcessCommandLine:
@@ -70,3 +108,71 @@ class TestProcessCommandLine:
         result = runner.invoke(process_command_line, ["-f", "nonexistent.rtf"])
         assert result.exit_code == 1
         assert "Error processing RTF file" in result.output
+
+    def test_process_all_files_integration(self) -> None:
+        """Test the integration of the 'process all files' functionality."""
+        with tempfile.TemporaryDirectory() as tmpdir:
+            tmpdir_path = Path(tmpdir)
+            rtf_file1 = tmpdir_path / "test1.rtf"
+            rtf_file2 = tmpdir_path / "test2.rtf"
+
+            # Create RTF files with proper RTF content
+            rtf_content1 = r"{\rtf1\ansi\deff0 Test RTF content 1}"
+            rtf_content2 = r"{\rtf1\ansi\deff0 Test RTF content 2}"
+            rtf_file1.write_text(rtf_content1)
+            rtf_file2.write_text(rtf_content2)
+
+            runner = CliRunner()
+
+            # First, test that both files would be found
+            from kp_analysis_toolkit.cli.common.file_selection import (
+                get_all_files_matching_pattern,
+            )
+
+            files = get_all_files_matching_pattern(tmpdir_path, "*.rtf")
+            expected_file_count = 2
+            assert len(files) == expected_file_count
+
+            # Mock user input to select "process all files" (option 3 with 2 files)
+            with patch("builtins.input", return_value="3"):
+                result = runner.invoke(process_command_line, ["-d", str(tmpdir_path)])
+
+            # Debug output in case of failure
+            print(f"Exit code: {result.exit_code}")
+            print(f"Output: {result.output}")
+            if result.exception:
+                print(f"Exception: {result.exception}")
+
+            # Should exit normally after processing all files
+            assert result.exit_code == 0
+
+            # Check that output files were created (they have timestamps in the name)
+            # Look for files matching the pattern test1_converted-*.txt and test2_converted-*.txt
+            output_files = list(tmpdir_path.glob("*_converted-*.txt"))
+
+            # List all files in directory for debugging
+            print(f"Files in directory: {list(tmpdir_path.iterdir())}")
+            print(f"Output files found: {output_files}")
+
+            assert len(output_files) == expected_file_count, (
+                f"Expected {expected_file_count} output files, found {len(output_files)}"
+            )
+
+            # Find the specific output files
+            test1_output = next(
+                (f for f in output_files if f.name.startswith("test1_converted-")),
+                None,
+            )
+            test2_output = next(
+                (f for f in output_files if f.name.startswith("test2_converted-")),
+                None,
+            )
+
+            assert test1_output is not None, "Expected test1 output file to exist"
+            assert test2_output is not None, "Expected test2 output file to exist"
+
+            # Verify content was processed
+            content1 = test1_output.read_text()
+            content2 = test2_output.read_text()
+            assert "Test RTF content 1" in content1
+            assert "Test RTF content 2" in content2
